reduce code refactor
